pyproject.toml
[tool.poetry]
name = "firstflask"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^2.0.1"
black = "^21.9b0"
Flask-SQLAlchemy = "^2.5.1"
Flask-WTF = "^0.15.1"
WTForm = "^1.0"
email-validator = "^1.1.3"
Flask-Bcrypt = "^0.7.1"
Flask-Login = "^0.5.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
